pyproject.toml
[project]
name = "reqify"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.73.1",
    "python-dotenv>=1.2.2",
]
